+++ CHANGELOG.md
@@ -12,7 +12,8 @@ x.y.z / 2014-06-dd
 * added methods `createEncoder()` and `createDecoder()`
 * changed method signature of `encode(version, payload)` to `encode(payload,version)`
 * changed method signature of `decode(version, base58str)` to `decode(base58str, version)`
-* 
+* added method `createValidator()`
+* changed method signature of `validate(version, base58str)` to `validate(base58str, version)`
 
 0.2.0 / 2014-03-10
 ------------------